9.1. A Simple Diffuse Material

+++ src/config.zig
@@ -1,2 +1,2 @@
-pub const float = f32;
+pub const float = f64;
 pub const inf = @import("std").math.inf(float);
